SSCS-8082 Moving common code into base class
